refactor: consolidating duplicated code
