+++ backend/src/sync/websocket.ts
@@ -4,11 +4,16 @@
  */
 
 import type { Auth } from '@/auth/elysia-plugin'
-import { user } from '@/db/auth-schema'
 import type { db as DbType } from '@/db/client'
 import { and, eq, gt } from 'drizzle-orm'
 import { Elysia, t } from 'elysia'
-import { syncChanges, syncDevices } from './schema'
+import { syncChanges } from './schema'
+import {
+  compareMigrationVersions,
+  getRequiredMigrationVersion,
+  updateMigrationVersionIfNewer,
+  upsertDevice,
+} from './utils'
 
 /**
  * Serialized change format for network transport
@@ -84,22 +89,6 @@ type ConnectedClient = {
   siteId: string
   migrationVersion?: string
   lastServerVersion: bigint
-}
-
-/**
- * Compare two migration versions
- */
-const compareMigrationVersions = (a: string | null, b: string | null): number => {
-  if (!a && !b) return 0
-  if (!a) return -1
-  if (!b) return 1
-
-  const getVersionNumber = (version: string): number => {
-    const match = version.match(/^(\d+)/)
-    return match ? parseInt(match[1], 10) : 0
-  }
-
-  return getVersionNumber(a) - getVersionNumber(b)
 }
 
 // Store connected clients by userId for broadcasting
@@ -193,13 +182,7 @@ export const createSyncWebSocketRoutes = (database: typeof DbType, auth: Auth) =
           const authUser = session.user
 
           // Check migration version
-          const currentUser = await database
-            .select({ syncMigrationVersion: user.syncMigrationVersion })
-            .from(user)
-            .where(eq(user.id, authUser.id))
-            .limit(1)
-
-          const requiredVersion = currentUser[0]?.syncMigrationVersion ?? null
+          const requiredVersion = await getRequiredMigrationVersion(database, authUser.id)
 
           if (requiredVersion && compareMigrationVersions(migrationVersion ?? null, requiredVersion) < 0) {
             ws.send(
@@ -213,25 +196,7 @@ export const createSyncWebSocketRoutes = (database: typeof DbType, auth: Auth) =
           }
 
           // Get or create device record
-          const existingDevice = await database
-            .select({ id: syncDevices.id })
-            .from(syncDevices)
-            .where(and(eq(syncDevices.userId, authUser.id), eq(syncDevices.siteId, siteId)))
-            .limit(1)
-
-          if (existingDevice.length > 0) {
-            await database
-              .update(syncDevices)
-              .set({ lastSeenAt: new Date(), migrationVersion })
-              .where(eq(syncDevices.id, existingDevice[0].id))
-          } else {
-            await database.insert(syncDevices).values({
-              userId: authUser.id,
-              siteId,
-              migrationVersion,
-              lastSeenAt: new Date(),
-            })
-          }
+          await upsertDevice(database, authUser.id, siteId, migrationVersion)
 
           // Get current server version
           const lastChange = await database
@@ -301,20 +266,8 @@ export const createSyncWebSocketRoutes = (database: typeof DbType, auth: Auth) =
 
           // Update migration version if newer
           if (client.migrationVersion) {
-            const currentUser = await database
-              .select({ syncMigrationVersion: user.syncMigrationVersion })
-              .from(user)
-              .where(eq(user.id, client.userId))
-              .limit(1)
-
-            const requiredVersion = currentUser[0]?.syncMigrationVersion ?? null
-
-            if (compareMigrationVersions(client.migrationVersion, requiredVersion) > 0) {
-              await database
-                .update(user)
-                .set({ syncMigrationVersion: client.migrationVersion })
-                .where(eq(user.id, client.userId))
-            }
+            const currentVersion = await getRequiredMigrationVersion(database, client.userId)
+            await updateMigrationVersionIfNewer(database, client.userId, client.migrationVersion, currentVersion)
           }
 
           // Insert changes
